Break up shopping list tests into multiple suites, partly for maintainability, partly because leaky state was causing tests to fail and it was due to a bug in testing library so I couldn't fix it
