+++ package.json
@@ -70,6 +70,7 @@
     "@storybook/node-logger": "^6.3.4",
     "@storybook/preset-create-react-app": "^3.2.0",
     "@storybook/react": "^6.3.4",
+    "canvas": "^2.8.0",
     "msw": "^0.30.0",
     "msw-storybook-addon": "^1.1.0",
     "universal-cookie": "^4.0.4"
